new booksearch.js file to add books to page from html, deleted old api get request folders, renamed any profile routes to the proper naming structure.
