+++ public/js/login.js
@@ -15,7 +15,7 @@ const loginFormHandler = async (event) => {
 
     if (response.ok) {
       // If successful, redirect the browser to the profile page
-      document.location.replace('/profile');
+      document.location.replace('/booksearch');
     } else {
       alert(response.statusText);
     }
@@ -25,10 +25,9 @@ const loginFormHandler = async (event) => {
 const signupFormHandler = async (event) => {
   event.preventDefault();
 
-  const name = document.querySelector('#name-signup').value.trim();
-  const email = document.querySelector('#email-signup').value.trim();
-  const password = document.querySelector('#password-signup').value.trim();
-
+  const name = document.querySelector('#name-signup').value?.trim();
+  const email = document.querySelector('#email-signup').value?.trim();
+  const password = document.querySelector('#password-signup').value?.trim();
   if (name && email && password) {
     const response = await fetch('/api/users', {
       method: 'POST',
@@ -37,7 +36,7 @@ const signupFormHandler = async (event) => {
     });
 
     if (response.ok) {
-      document.location.replace('/profile');
+      document.location.replace('/booksearch');
     } else {
       alert(response.statusText);
     }
