+++ views/booksearch.handlebars
@@ -4,9 +4,14 @@
   </div>
 </div>
 
+   <form class="d-flex" role="search">
+    <input id="searchInput" class="form-control me-2" type="search" placeholder="Search" aria-label="Search">
+    <button id="searchBtn" class="btn btn-outline-dark" type="submit">Search</button>
+  </form>
+
 <div class="row mt-4">
   <div class="col-md-6">
-    <h3>Top Books</h3>
+    <h3>My Books</h3>
 
     <form class="form new-book-form">
       <div class="form-group">
@@ -45,4 +50,4 @@
   {{/if}}
 </div>
 
-<script src="./js/profile.js"></script>
+<script src="./js/booksearch.js"></script>
