+++ views/homepage.handlebars
@@ -1,3 +1,5 @@
+
+ 
 <!-- Loop over each object in the books array -->
 {{#each books as |book| }}
 <div class="row mb-4 book">
@@ -22,3 +24,5 @@
   </div>
 </div>
 {{/each}}
+
+<script src="/js/booksearch.js"></script>
