+++ views/login.handlebars
@@ -19,16 +19,16 @@
 
 <form class="row g-3 form signup-form">
     <div class="input-group mb-3">
-      <span class="input-group-text" id="name-signup">Name</span>
-      <input type="text" class="form-control" placeholder="Please Enter Name Here" aria-label="Name"
+      <label class="input-group-text">Name</label>
+      <input type="text" class="form-control" id="name-signup" placeholder="Please Enter Name Here" aria-label="Name"
         aria-describedby="basic-addon1">
     </div>
   <div class="col-auto">
-    <label for="staticEmail2" class="visually-hidden">Email</label>
-    <input type="text" class="form-control-plaintext" id="email-signup" value="Email">
+    <label for="staticEmail2">Email</label>
+    <input type="email" class="form-control" id="email-signup" placeholder="Email">
   </div>
   <div class="col-auto">
-    <label for="inputPassword2" class="visually-hidden">Password</label>
+    <label for="inputPassword2">Password</label>
     <input type="password" class="form-control" id="password-signup" placeholder="Password">
   </div>
   <div class="col-auto">
